Add `spmm` bf16 support (#269)
* Add spmm bf16 support and add ut

* Add bf16 support for spspmm

* Disable bf16 test before torch_scatter 2.0.9

* Refactor version compare

* Update test/utils.py

+++ test/test_spspmm.py
@@ -35,7 +35,7 @@ def test_sparse_tensor_spspmm(dtype, device):
         ], dtype=dtype, device=device),
     )
 
-    expected = torch.eye(10, dtype=dtype, device=device)
+    expected = torch.eye(10, device=device).to(dtype)
 
     out = x @ x.to_dense().t()
     assert torch.allclose(out, expected, atol=1e-2)
